services/appointment-service/requirements.txt
Flask==2.1.1
Flask-JWT-Extended==4.3.1
pymongo==3.12.1
bson==0.5.10
python-dotenv==0.19.2
APScheduler==3.10.4
twilio==7.0.0
Werkzeug==2.1.1
